# data/products.go
package data

import (
	"encoding/json"
	"fmt"
	"io"
	"regexp"
	"time"

	"github.com/go-playground/validator/v10"
)

// Product defines the structure for an API Product
// swagger:model
type Product struct {
	// the id for this user
	// 
	// required: true
	// min: 1
	ID				int `json:"id"`
	Name			string `json:"name" validate:"required"`
	Description		string `json:"description"`
	Price			float32 `json:"price" validate:"gt=0"`
	SKU				string `json:"sku" validate:"required,sku"`
	CreatedOn		string `json:"created_at"`
	UpdatedOn		string `json:"updated_at"`
	DeletedOn		string `json:"-"`
}


func (p *Product) FromJson(r io.Reader) error {
	e := json.NewDecoder(r)
	return e.Decode(p)
}

func (p *Product) Validate() error  {
	validate := validator.New()
	validate.RegisterValidation("sku", validateSKU)
	return validate.Struct(p)
}

func validateSKU(fl validator.FieldLevel) bool  {
	re := regexp.MustCompile(`[a-z]+-[a-z]+-[a-z]+`)
	matches := re.FindAllString(fl.Field().String(), -1)
	return len(matches) == 1
}

type Products []*Product

func (p *Products) ToJson(w io.Writer) error {
	e := json.NewEncoder(w)
	return e.Encode(p)
}

func GetProducts() Products  {
	return productList
}

func AddProduct(p *Product)  {
	p.ID = getNextID()
	productList = append(productList, p)
}

func UpdateProduct(id int, p *Product)  error {
	_, pos, err := findProduct(id)
	if err != nil {
		return err
	}
	p.ID = id
	productList[pos] = p
	return nil
}

func DeleteProduct(id int, p *Product)  error {
	_, pos, err := findProduct(id)
	if err != nil {
		return err
	}
	p.ID = id
	productList[pos] = nil
	return nil
}

var ErrProductNotFound = fmt.Errorf("Product not found")

func findProduct(id int) (*Product, int, error) {
	for i, p := range productList {
		if p.ID == id {
			return p, i, nil
		}
	}
	return nil, -1, ErrProductNotFound
}


func getNextID() int {
	lp := productList[len(productList) - 1]
	return lp.ID + 1
}

var productList = []*Product{
	{
		ID: 1,
		Name: "Latte",
		Description: "Frostry",
		Price: 2.45,
		SKU: "abcd123",
		CreatedOn: time.Now().UTC().String(),
		UpdatedOn: time.Now().UTC().String(),
	},
	{
		ID: 2,
		Name: "Latte",
		Description: "Frostry",
		Price: 2.45,
		SKU: "abcd123",
		CreatedOn: time.Now().UTC().String(),
		UpdatedOn: time.Now().UTC().String(),
	},	
}